requirements.txt
pandas
numpy
kafka-python
pyspark==3.5.0
redis
pymongo
streamlit
plotly
python-dotenv
requests
streamlit-extras
streamlit-autorefresh
influxdb
